ADM-359.1 rename features as context
