+++ frontend/__tests__/src/features/boardSlice.test.tsx
@@ -1,4 +1,4 @@
-import boardReducer, { changeBoardVerifyState } from '@src/features/board/boardSlice'
+import boardReducer, { changeBoardVerifyState } from '@src/context/board/boardSlice'
 
 describe('board reducer', () => {
   it('should return false when handle initial state', () => {
