+++ frontend/__tests__/src/utils/setupStoreUtil.tsx
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit'
-import { stepperSlice } from '@src/features/stepper/StepperSlice'
-import { configSlice } from '@src/features/config/configSlice'
-import { boardSlice } from '@src/features/board/boardSlice'
+import { stepperSlice } from '@src/context/stepper/StepperSlice'
+import { configSlice } from '@src/context/config/configSlice'
+import { boardSlice } from '@src/context/board/boardSlice'
 
 export const setupStore = () => {
   return configureStore({
