+++ frontend/src/components/Metrics/ConfigStep/DateRangePicker/index.tsx
@@ -7,8 +7,8 @@ import { TextField } from '@mui/material'
 import { datePickerPropsStyles, DateRangeBox } from '@src/components/Metrics/ConfigStep/DateRangePicker/style'
 import { DATE_RANGE, SELECT_OR_WRITE_DATE } from '@src/constants'
 import { useAppDispatch, useAppSelector } from '@src/hooks/useAppDispatch'
-import { changeBoardVerifyState } from '@src/features/board/boardSlice'
-import { selectDateRange, updateDateRange } from '@src/features/config/configSlice'
+import { changeBoardVerifyState } from '@src/context/board/boardSlice'
+import { selectDateRange, updateDateRange } from '@src/context/config/configSlice'
 
 export const DateRangePicker = () => {
   const dispatch = useAppDispatch()
