+++ frontend/src/store.ts
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit'
-import stepperReducer from './features/stepper/StepperSlice'
-import configReducer from './features/config/configSlice'
-import boardReducer from './features/board/boardSlice'
-import jiraVerifyResponseReducer from './features/board/jiraVerifyResponse/jiraVerifyResponseSlice'
+import stepperReducer from './context/stepper/StepperSlice'
+import configReducer from './context/config/configSlice'
+import boardReducer from './context/board/boardSlice'
+import jiraVerifyResponseReducer from './context/board/jiraVerifyResponse/jiraVerifyResponseSlice'
 export const store = configureStore({
   reducer: {
     stepper: stepperReducer,
